+++ controllers/service_controller_test.go
@@ -32,164 +32,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/event"
 )
 
-func TestShouldWatchSrv(t *testing.T) {
-	a := assert.New(t)
-	cases := []struct {
-		srv    *corev1.Service
-		cli    client.Client
-		u      *Utils
-		expRes bool
-	}{
-		{
-			srv: &corev1.Service{
-				ObjectMeta: metav1.ObjectMeta{Name: "srv-name", Namespace: "ns-name"},
-				Spec: corev1.ServiceSpec{
-					Type: corev1.ServiceTypeClusterIP,
-				},
-			},
-		},
-		{
-			srv: &corev1.Service{
-				ObjectMeta: metav1.ObjectMeta{Name: "srv-name", Namespace: "ns-name"},
-				Spec: corev1.ServiceSpec{
-					Type: corev1.ServiceTypeLoadBalancer,
-				},
-				Status: corev1.ServiceStatus{
-					LoadBalancer: corev1.LoadBalancerStatus{
-						Ingress: []corev1.LoadBalancerIngress{},
-					},
-				},
-			},
-		},
-		{
-			u: &Utils{
-				AllowedAnnotations: []string{"yes"},
-			},
-			srv: &corev1.Service{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        "srv-name",
-					Namespace:   "ns-name",
-					Annotations: map[string]string{"no": "no"},
-				},
-				Spec: corev1.ServiceSpec{
-					Type: corev1.ServiceTypeLoadBalancer,
-				},
-				Status: corev1.ServiceStatus{
-					LoadBalancer: corev1.LoadBalancerStatus{
-						Ingress: []corev1.LoadBalancerIngress{
-							{IP: "10.10.10.10"},
-						},
-					},
-				},
-			},
-		},
-		{
-			cli: fakecli.NewFakeClientWithScheme(&runtime.Scheme{}),
-			u: &Utils{
-				AllowedAnnotations: []string{"yes"},
-			},
-			srv: &corev1.Service{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        "srv-name",
-					Namespace:   "ns-name",
-					Annotations: map[string]string{"yes": "no"},
-				},
-				Spec: corev1.ServiceSpec{
-					Type: corev1.ServiceTypeLoadBalancer,
-				},
-				Status: corev1.ServiceStatus{
-					LoadBalancer: corev1.LoadBalancerStatus{
-						Ingress: []corev1.LoadBalancerIngress{
-							{IP: "10.10.10.10"},
-						},
-					},
-				},
-			},
-		},
-		{
-			cli: func() client.Client {
-				scheme := runtime.NewScheme()
-				corev1.AddToScheme(scheme)
-				return fakecli.NewFakeClientWithScheme(scheme, &corev1.Namespace{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "ns-name",
-					},
-				})
-			}(),
-			u: &Utils{
-				AllowedAnnotations: []string{"yes"},
-			},
-			srv: &corev1.Service{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        "srv-name",
-					Namespace:   "ns-name",
-					Annotations: map[string]string{"yes": "no"},
-				},
-				Spec: corev1.ServiceSpec{
-					Type: corev1.ServiceTypeLoadBalancer,
-				},
-				Status: corev1.ServiceStatus{
-					LoadBalancer: corev1.LoadBalancerStatus{
-						Ingress: []corev1.LoadBalancerIngress{
-							{IP: "10.10.10.10"},
-						},
-					},
-				},
-			},
-		},
-		{
-			cli: func() client.Client {
-				scheme := runtime.NewScheme()
-				corev1.AddToScheme(scheme)
-				return fakecli.NewFakeClientWithScheme(scheme, &corev1.Namespace{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:   "ns-name",
-						Labels: map[string]string{string(types.AllowedKey): "whatever"},
-					},
-				})
-			}(),
-			u: &Utils{
-				AllowedAnnotations: []string{"yes"},
-				CurrentNsPolicy:    types.AllowList,
-			},
-			srv: &corev1.Service{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        "srv-name",
-					Namespace:   "ns-name",
-					Annotations: map[string]string{"yes": "no"},
-				},
-				Spec: corev1.ServiceSpec{
-					Type: corev1.ServiceTypeLoadBalancer,
-				},
-				Status: corev1.ServiceStatus{
-					LoadBalancer: corev1.LoadBalancerStatus{
-						Ingress: []corev1.LoadBalancerIngress{
-							{IP: "10.10.10.10"},
-						},
-					},
-				},
-			},
-			expRes: true,
-		},
-	}
-	failed := func(i int) {
-		a.FailNow(fmt.Sprintf("case %d failed", i))
-	}
-	for i, currCase := range cases {
-		s := &ServiceReconciler{
-			Client:        currCase.cli,
-			Log:           ctrl.Log.WithName("test"),
-			Utils:         currCase.u,
-			cacheSrvWatch: map[string]bool{},
-		}
-
-		res := s.shouldWatchSrv(currCase.srv)
-		if !a.Equal(currCase.expRes, res) {
-			failed(i)
-		}
-	}
-}
-
 func TestSrvUpdatePredicate(t *testing.T) {
 	a := assert.New(t)
 	cli := func() client.Client {
